break out components in my-calendar
